requirements.txt
Django==2.2
djangorestframework==3.8.2
psycopg2==2.7.3.2
